solana_rpc_client_api

Module response

Re-exports

pub use solana_rpc_client_types::response::transaction;

Structs

EncodedTransactionWithStatusMeta

FeeCalculator

FeeRateGovernor

ParsedAccount

ParsedInstruction

ProcessedSignatureResult

Response

Reward

RpcAccountBalance

RpcApiVersion

RpcBlockCommitment

RpcBlockProduction

RpcBlockProductionRange

RpcBlockUpdate

RpcBlockhash

RpcBlockhashFeeCalculator

RpcConfirmedTransactionStatusWithSignature

RpcContactInfo

RpcFeeCalculator

RpcFeeRateGovernor

RpcIdentity

RpcInflationGovernor

RpcInflationRate

RpcInflationReward

RpcKeyedAccount

RpcLogsResponse

RpcPerfSample

RpcPrioritizationFee

RpcResponseContext

RpcSignatureConfirmation

RpcSimulateTransactionResult

RpcSnapshotSlotInfo

RpcStorageTurn

RpcSupply

RpcTokenAccountBalance

RpcVersionInfo

RpcVote

RpcVoteAccountInfo

RpcVoteAccountStatus

SlotInfo

SlotTransactionStats

TransactionParsedAccount

UiAccount

A duplicate representation of an Account for pretty JSON serialization

UiAccountsList

UiCompiledInstruction

A duplicate representation of a CompiledInstruction for pretty JSON serialization

UiConfirmedBlock

UiInnerInstructions

UiLoadedAddresses

A duplicate representation of LoadedAddresses

UiPartiallyDecodedInstruction

A partially decoded CompiledInstruction that includes explicit account addresses

UiTokenAmount

UiTransactionError

UiTransactionReturnData

UiTransactionStatusMeta

A duplicate representation of TransactionStatusMeta with err field

UiTransactionTokenBalance

Enums

EncodedTransaction

OptionSerializer

OptionalContext

Wrapper for rpc return types of methods that provide responses both with and without context. Main purpose of this is to fix methods that lack context information in their return type, without breaking backwards compatibility.

ReceivedSignatureResult

RewardType

RpcBlockUpdateError

RpcSignatureResult

SlotUpdate

StakeActivationState

TransactionBinaryEncoding

TransactionConfirmationStatus

TransactionError

Reasons a transaction might be rejected.

UiAccountData

UiAccountEncoding

UiInstruction

A duplicate representation of an Instruction for pretty JSON serialization

UiParsedInstruction

UiReturnDataEncoding

Value

Represents any valid JSON value.

Type Aliases

Rewards

RpcLeaderSchedule

Map of leader base58 identity pubkeys to the slot indices relative to the first epoch slot

RpcResult

TransactionResult