solana_rpc_client_api

Module response

Structs

DeprecatedRpcFees

Fees

ProcessedSignatureResult

Response

RpcAccountBalance

RpcApiVersion

RpcBlockCommitment

RpcBlockProduction

RpcBlockProductionRange

RpcBlockUpdate

RpcBlockhash

RpcBlockhashFeeCalculator

RpcConfirmedTransactionStatusWithSignature

RpcContactInfo

RpcFeeCalculator

RpcFeeRateGovernor

RpcFees

RpcIdentity

RpcInflationGovernor

RpcInflationRate

RpcInflationReward

RpcKeyedAccount

RpcLogsResponse

RpcPerfSample

RpcPrioritizationFee

RpcResponseContext

RpcSignatureConfirmation

RpcSimulateTransactionResult

RpcSnapshotSlotInfo

RpcStakeActivation

RpcStorageTurn

RpcSupply

RpcTokenAccountBalance

RpcVersionInfo

RpcVote

RpcVoteAccountInfo

RpcVoteAccountStatus

SlotInfo

SlotTransactionStats

Enums

OptionalContext

Wrapper for rpc return types of methods that provide responses both with and without context. Main purpose of this is to fix methods that lack context information in their return type, without breaking backwards compatibility.

ReceivedSignatureResult

RpcBlockUpdateError

RpcSignatureResult

SlotUpdate

StakeActivationState

Type Aliases

RpcLeaderSchedule

Map of leader base58 identity pubkeys to the slot indices relative to the first epoch slot

RpcResult